refactor out seed functions into new file
